Create codsoft task 2
Python provides a standard GUI framework Tkinter which is used to develop fast and easy GUI applications. Here, I developed a simple multiple-choice Cricket based quiz in python with GUI.I created a multiple choice quiz in Python with Tkinter.

This quiz game uses a function, a block of code with at name that performs a specific task. A function allows us to use the same code several times, without having to type everything each time. Python has a lot of built-in functions, but it also allows to create our own functions. The score is stored in a variable during the game. Once all the questions have been answered, the game ends.

For each correct answer 1 point will be awarded. At the end of the game, the program will reveal the final score.